testing added

+++ backend/services/PostService.js
@@ -1,3 +1,4 @@
+const { isValidObjectId } = require("mongoose")
 const Post = require("../models/Post")
 
 class PostService{
@@ -26,28 +27,64 @@ class PostService{
     }
 
     getById = async(id) => {
-        const post = await Post.findById(id)
-        return post
+        try {
+            if(!isValidObjectId(id)) throw new Error('id received not is a mongoID valid')
+            const post = await Post.findById(id)
+            return post   
+        } catch (error) {
+            return {
+                error: true,
+                msgError: error
+            }
+        }
     }
 
     createOne = async({username, title, image}) => {
-        const post = new Post({
-            username,
-            title,
-            image
-        })
-        await post.save()
-        return post
+        try {
+            if(username == '' || username.includes(' ')) throw new Error('username should not void or content a spaces.')
+            if(typeof(username) != 'string') throw new Error('username must be a string type.')
+            if(title != undefined && typeof(title)!='string') throw new Error('title must be undefined or string type.')
+            if(typeof(image)!='string') throw new Error('image must be a string type.')
+            if(!(image.includes('https://', 0) || image.includes('http://', 0))) throw new Error('image must be a valid url.')
+            const post = new Post({
+                username,
+                title,
+                image
+            })
+            await post.save()
+            return post   
+        } catch (error) {
+            return {
+                error: true,
+                msgError: error
+            }
+        }
     }
 
     addLike = async(id) => {
-        const post = await Post.findByIdAndUpdate(id, {likes: post.likes}, {new: true})
-        return post
+        try {
+            if(!isValidObjectId(id)) throw new Error('id must be a MongoID type.')
+            const post = await Post.findByIdAndUpdate(id, {likes: post.likes}, {new: true})
+            return post   
+        } catch (error) {
+            return{
+                error: true,
+                msgError: error
+            }
+        }
     }
 
     deleteById = async(id) => {
-        const post = await Post.findByIdAndDelete(id)
-        return post
+        try {
+            if(!isValidObjectId(id)) throw new Error('id received not is a mongoID valid')
+            const post = await Post.findByIdAndDelete(id)
+            return post
+        } catch (error) {
+            return{
+                error: true,
+                msgError: error
+            }
+        }
     }
 
     changeTitleService = async(id, newTitle) => {
